make all primary keys biginteger
- make all primary keys biginteger
- change primary key of telegram subscriptions
- add tests for create_subscription

+++ migrations/2020-02-22-080419_create_feeds/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE feeds(
-   id SERIAL PRIMARY KEY,
+   id BIGSERIAL PRIMARY KEY,
    title TEXT,
    link TEXT NOT NULL UNIQUE,
    error TEXT,
@@ -13,8 +13,8 @@ ALTER TABLE feeds
       ADD CONSTRAINT feed_link_size CHECK (char_length(link) > 0);
 
 create TABLE feed_items(
-   id SERIAL PRIMARY KEY,
-   feed_id INTEGER NOT NULL references feeds(id) ON DELETE CASCADE,
+   id BIGSERIAL PRIMARY KEY,
+   feed_id BIGINT NOT NULL references feeds(id) ON DELETE CASCADE,
    title TEXT,
    description TEXT,
    link TEXT,
